Revert "Add a main util module that allows us to use the validator from other modules."
This reverts commit a8f4a30895e855c6c2c543c4dd3b78cff345f29c.
