+++ package.json
@@ -22,7 +22,7 @@
     "validate-office-addin": "./app/index.js"
   },
   "author": "OfficeDev",
-  "main": "./app/util.js",
+  "main": "",
   "keywords": [
     "office",
     "validation",
@@ -41,6 +41,7 @@
     "applicationinsights": "^0.19.0",
     "chalk": "^1.1.3",
     "commander": "^2.9.0",
+    "node-status": "^1.0.0",
     "request": "^2.81.0",
     "request-promise": "^4.2.0"
   },
@@ -51,7 +52,7 @@
     "@types/chalk": "^0.4.31",
     "@types/commander": "^2.3.31",
     "@types/mocha": "^2.2.40",
-    "@types/request": "^0.0.45",
+    "@types/request": "^0.0.42",
     "@types/request-promise": "^4.1.33",
     "chai": "^3.5.0",
     "chai-as-promised": "^6.0.0",
